Implement a ResumableCommitLogReader
Adds a new Single Source of Reading Truth for CommitLogSegments.
This resumable reader has a simpler API than our existing CommitLogReader
and is primarily intended to be externally user-facing.

Primary tradeoffs of this implementation are 1) Requiring init of a new
ResumableCommitLogReader object on each CLS read, 2) taking on ownership of
some complexity in SegmentReader (iterator resumability and RAR reinit)
in order to provide a very simple API to external CDC users, and 3) splitting
out the "Reading a CommitLog" API into both part in CommitLogReader and part
in ResumableCommitLogReader.

For now, for API backwards compatibility, the CommitLogReader APIs are
unchanged and ResumableCommitLogReader's API is purely simplified and
additive.