Refactor commit log testing and cdc testing to support resumable testing
The original CDC CommitLogReadHandler for test only handled a boolean state
of whether or not a cdc mutation was seen or not. This opens that up to allow
for a tracking of a MutationIdentifier to determine if the specific mutation
we're seeing on replay is a duplicate and if so, how many times we've seen it.
This will play into resumable clreader testing in subsequent commits (some
risk of yagni here, but have a general framework of how I want to test this
in my head. We'll see.)